mv vqvae.py vqvae_tester.py from  HiFi-Codec-24k-320d to HiFi-Codec
