+++ HiFi-Codec/vqvae_tester.py
@@ -14,17 +14,16 @@ from vqvae import VQVAE
 
 class VqvaeTester(nn.Module):
 
-    def __init__(self, hp):
+    def __init__(self, config_path, model_path, sample_rate=24000):
         super().__init__()
-        self.hp = hp
-        self.vqvae = VQVAE(hp.config_path, hp.model_path, with_encoder=True)
-        self.sample_rate = self.hp.sample_rate
+        self.vqvae = VQVAE(config_path, model_path, with_encoder=True)
+        self.sample_rate = sample_rate
 
     @torch.no_grad()
     def forward(self, wav_path):
         wav, sr = sf.read(wav_path)
         if sr != self.sample_rate:
-            wav = librosa.resample(wav, orig_sr=sr, target_sr=self.hp.sample_rate)
+            wav = librosa.resample(wav, orig_sr=sr, target_sr=self.sample_rate)
         fid = os.path.basename(wav_path)[:-4]
         wav = normalize(wav) * 0.95
         wav = torch.FloatTensor(wav).unsqueeze(0)
@@ -37,7 +36,7 @@ class VqvaeTester(nn.Module):
     def vq(self, wav_path):
         wav, sr = sf.read(wav_path)
         if sr != self.sample_rate:
-            wav = librosa.resample(wav, orig_sr=sr, target_sr=self.hp.sample_rate)
+            wav = librosa.resample(wav, orig_sr=sr, target_sr=self.sample_rate)
         fid = os.path.basename(wav_path)[:-4]
         wav = normalize(wav) * 0.95
         wav = torch.FloatTensor(wav).unsqueeze(0)
